feeders/digitraffic-road/pyproject.toml
[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "digitraffic-road"
version = "0.1.0"
description = "Finnish road traffic data bridge (sensors, traffic messages, maintenance tracking) via Digitraffic MQTT"
authors = ["Clemens Vasters <clemensv@microsoft.com>"]
packages = [{ include = "digitraffic_road" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
paho-mqtt = ">=2.1.0"
confluent-kafka = ">=2.5.3"
cloudevents = ">=1.11.0,<2.0.0"
requests = ">=2.31.0"
dataclasses-json = ">=0.6.7"
digitraffic_road_producer_data = { path = "digitraffic_road_producer/digitraffic_road_producer_data" }
digitraffic_road_producer_kafka_producer = { path = "digitraffic_road_producer/digitraffic_road_producer_kafka_producer" }

[tool.poetry.dev-dependencies]
pytest = ">=8.3.3"
testcontainers = ">=4.8.1"

[tool.poetry.scripts]
digitraffic-road = "digitraffic_road:main"
